TimeLine chart

+++ stall/BK-SFCS-Stall/controller/sales.h
@@ -26,7 +26,7 @@ public slots:
     void setStartRange(double pos);
     void setEndRange(double pos);
     unsigned drawQuantityBarGraph(QAbstractSeries *series);
-    int drawTimeLineGraph(QAbstractSeries *series1, QAbstractSeries *series2);
+    unsigned long long drawTimeLineGraph(QAbstractSeries *series1, QAbstractSeries *series2);
     unsigned drawRejectedBarGraph(QAbstractSeries *series);
     void advancedExport();
 signals:
